Added package structure which puts everything under src, then ericjb then aiprojects.  Individual projects will each be in their own package under aiprojects.
Completed the online Python Tutorial, the results are the shopSmart.py and buyLotsOfFruit.py files in their respective packages.

Added Quicksort under the sort package which pivots based on the first entry in the list.

Added foreach, a very simple program as part of the tutorial

Added a more functionally oriented listcomp2, which demonstrates basic use of the map(), filter(), and reduce() functions.

Added test for personal experimental programs