Working on fourCornersProblem:
Thoughts on what's going wrong:
The frontierSet is becoming empty before all state's have been tried.  However, all positions have been visited.

getSuccessors needs to be able to return configurations which may have been expanded, and visited before.

It would be nice if the frontier and exploredset could empty once a corner was reached, but I don't think this is possible.

An alternative might be to consider states new based on how many corners its visited.  This would require the frontier and exploredset to record this information somehow.